[Core] Move Administrator to it’s own namespace
The Core Component is about more then only the Administrator User,
to integrate more functionality in the Core component this change makes
a better organization.
