+++ src/Bundle/CoreBundle/Resources/config/Doctrine/Administrator/Administrator.orm.xml
@@ -4,7 +4,7 @@
     xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
         http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="ParkManager\Component\Core\Model\Administrator" table="administrator">
+    <entity name="ParkManager\Component\Core\Model\Administrator\Administrator" table="administrator">
         <field name="firstName" column="first_name" type="string" />
         <field name="lastName" column="last_name" type="string" />
     </entity>
