+++ src/Bundle/CoreBundle/Resources/config/services/administrator.php
@@ -35,7 +35,7 @@ use ParkManager\Bundle\UserBundle\ReadModel\DbalUserFinder;
 use ParkManager\Bundle\UserBundle\Security\DbalUserProvider;
 use ParkManager\Bundle\UserBundle\Security\FormAuthenticator;
 use ParkManager\Bundle\UserBundle\Service\PasswordResetSwiftMailer;
-use ParkManager\Component\Core\Model\Handler\RegisterAdministratorHandler;
+use ParkManager\Component\Core\Model\Administrator\Handler\RegisterAdministratorHandler;
 use ParkManager\Component\Model\Event\EventEmitter;
 use ParkManager\Component\ServiceBus\QueryBus;
 use ParkManager\Component\User\Canonicalizer\SimpleEmailCanonicalizer;
