+++ src/Component/Core/Model/Administrator/AdministratorRepository.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-namespace ParkManager\Component\Core\Model;
+namespace ParkManager\Component\Core\Model\Administrator;
 
 use ParkManager\Component\User\Model\UserCollection;
 use ParkManager\Component\User\Model\UserId;
