+++ src/Component/Core/Model/Administrator/Event/AdministratorWasRegistered.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-namespace ParkManager\Component\Core\Model\Event;
+namespace ParkManager\Component\Core\Model\Administrator\Event;
 
 use ParkManager\Component\Model\Event\DomainEvent;
 use ParkManager\Component\User\Model\UserId;
