+++ src/Component/Core/Tests/Model/Administrator/Event/AdministratorNameWasChangedTest.php
@@ -12,9 +12,9 @@ declare(strict_types=1);
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-namespace ParkManager\Component\Core\Tests\Model\Event;
+namespace ParkManager\Component\Core\Tests\Model\Administrator\Event;
 
-use ParkManager\Component\Core\Model\Event\AdministratorNameWasChanged;
+use ParkManager\Component\Core\Model\Administrator\Event\AdministratorNameWasChanged;
 use ParkManager\Component\User\Model\UserId;
 use PHPUnit\Framework\TestCase;
