+++ src/Component/Core/Tests/Model/Administrator/Event/AdministratorWasRegisteredTest.php
@@ -12,9 +12,9 @@ declare(strict_types=1);
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-namespace ParkManager\Component\Core\Tests\Model\Event;
+namespace ParkManager\Component\Core\Tests\Model\Administrator\Event;
 
-use ParkManager\Component\Core\Model\Event\AdministratorWasRegistered;
+use ParkManager\Component\Core\Model\Administrator\Event\AdministratorWasRegistered;
 use ParkManager\Component\User\Model\UserId;
 use PHPUnit\Framework\TestCase;
