+++ src/Component/Core/Tests/Model/Administrator/Handler/RegisterUserHandlerTest.php
@@ -12,12 +12,12 @@ declare(strict_types=1);
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-namespace ParkManager\Component\Core\Tests\Model\Handler;
+namespace ParkManager\Component\Core\Tests\Model\Administrator\Handler;
 
-use ParkManager\Component\Core\Exception\AdministratorEmailAddressAlreadyInUse;
-use ParkManager\Component\Core\Model\Administrator;
-use ParkManager\Component\Core\Model\Command\RegisterAdministrator;
-use ParkManager\Component\Core\Model\Handler\RegisterAdministratorHandler;
+use ParkManager\Component\Core\Model\Administrator\Administrator;
+use ParkManager\Component\Core\Model\Administrator\Command\RegisterAdministrator;
+use ParkManager\Component\Core\Model\Administrator\Exception\AdministratorEmailAddressAlreadyInUse;
+use ParkManager\Component\Core\Model\Administrator\Handler\RegisterAdministratorHandler;
 use ParkManager\Component\User\Model\User;
 use ParkManager\Component\User\Model\UserCollection;
 use ParkManager\Component\User\Model\UserId;
